% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debtnetworkvisualizer-package.R
\docType{package}
\name{debtnetworkvisualizer-package}
\alias{debtnetworkvisualizer}
\alias{debtnetworkvisualizer-package}
\title{debtnetworkvisualizer: Debt Network Visualizer}
\description{
Visualize global debt networks based on World Bank International Debt Statistics (IDS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-intelligence/app-debt-network-visualizer}
  \item Report bugs at \url{https://github.com/tidy-intelligence/app-debt-network-visualizer/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} [copyright holder]

}
\keyword{internal}
